Cut resources by a bit more than half
We had overprovisioned to ensure the site doesn't fall over
when I was traveling. It hasn't, and this cuts it to a more
reasonable size.

We still need to tune the evaluator, as the sample submission
doesn't run to completion yet.

Also deploys both staging and prod even if only common.yaml is
changed. This was the cause of the launch deploy not triggering
a prod deployment.